Added support for starting the properties editor editing when hitting ENTER. Code cleaning.
git-svn-id: https://svn.ctdp.net/svn/rfdynhud/trunk@382 1c0d6ce8-a1e3-b74c-a49d-f790e2421cd5

+++ java/src/net/ctdp/rfdynhud/gamedata/FuelUsageRecorder.java
@@ -249,7 +249,7 @@ public class FuelUsageRecorder implements ScoringInfo.ScoringInfoUpdateListener
             
             if ( stintLength >= 2 )
             {
-                average = relevantFuel / (float)( fuelRelevantLaps );
+                average = relevantFuel / fuelRelevantLaps;
                 
                 onValuesUpdated( gameData, fuelRelevantLaps, relevantFuel, lastLap, average );
             }
